feat(Finalizado): :white_check_mark: Arreglos finalizados

+++ src/main/java/domain/service/UserService.java
@@ -4,5 +4,6 @@ import domain.entity.User;
 
 public interface UserService {
     void createUser(User user);
-    User findUserById(Long id);
+
+    User findUserById(int id);
 }
